feat: fractional word spacing for justified text
Use 16.8 fixed-point for word positions to eliminate integer
truncation in justified text spacing. This prevents uneven
word gaps caused by integer division rounding.

Changes:
- TextBlock.wordXpos: int16_t -> int32_t (fixed-point)
- ParsedText: calculate spacing in fixed-point, snap only at render
- Bump SECTION_FILE_VERSION 18->19 to invalidate old caches

+++ lib/Epub/Epub/blocks/TextBlock.h
@@ -13,7 +13,7 @@
 class TextBlock final : public Block {
  private:
   std::vector<std::string> words;
-  std::vector<int16_t> wordXpos;
+  std::vector<int32_t> wordXpos;
   std::vector<EpdFontFamily::Style> wordStyles;
   BlockStyle blockStyle;
